requirements.txt
fastapi
uvicorn
pydantic
asyncpg
sqlalchemy[asyncio]
redis
python-dotenv
openai 